Fix warnings generated by GCC 4.3.3 under Linux x86-64.
git-svn-id: https://edk2-buildtools.svn.sourceforge.net/svnroot/edk2-buildtools/trunk/BaseTools@1676 c2973edb-eda0-4c78-bc6a-9341b269661f

+++ Source/C/GenFv/GenFv.c
@@ -691,12 +691,20 @@ Returns:
     fprintf (FpFile, "\n");
     if (mFvDataInfo.BootBaseAddress != 0) {
       fprintf (FpFile, EFI_FV_BOOT_DRIVER_BASE_ADDRESS_STRING);
-      fprintf (FpFile, " = 0x%llx\n", mFvDataInfo.BootBaseAddress);
+      fprintf (
+        FpFile,
+        " = 0x%llx\n",
+        (unsigned long long)mFvDataInfo.BootBaseAddress
+        );
       DebugMsg (NULL, 0, 9, "Updated boot driver base address", "%s = 0x%llx", EFI_FV_RUNTIME_DRIVER_BASE_ADDRESS_STRING, (unsigned long long) mFvDataInfo.BootBaseAddress);
     }
     if (mFvDataInfo.RuntimeBaseAddress != 0) {
       fprintf (FpFile, EFI_FV_RUNTIME_DRIVER_BASE_ADDRESS_STRING);
-      fprintf (FpFile, " = 0x%llx\n", mFvDataInfo.RuntimeBaseAddress);
+      fprintf (
+        FpFile,
+        " = 0x%llx\n",
+        (unsigned long long)mFvDataInfo.RuntimeBaseAddress
+        );
       DebugMsg (NULL, 0, 9, "Updated runtime driver base address", "%s = 0x%llx", EFI_FV_RUNTIME_DRIVER_BASE_ADDRESS_STRING, (unsigned long long) mFvDataInfo.RuntimeBaseAddress);
     }
     fclose (FpFile);
